requirements.txt
fastapi
uvicorn
python-multipart
pyzbar
pillow